mm-drivers: hw_fence: avoid refcount leak with signaled dma-fences
When importing an already signaled dma-fence, the HW Fence driver may
create and destroy a backing hw-fence and then create a new hw-fence as
the backing hw-fence is destroyed.
Update the implementation to avoid unnecessary creation and destruction
of a fence and use a dummy client ID instead of the importing client ID
so that the driver does not incorrectly treat this as importing a fence
that the client will later signal.

Change-Id: Icf873112971c1abf2988a65ba880c31c920aa01a
Signed-off-by: Grace An <gracan@qti.qualcomm.com>