mm-driver: hfi_core: Add patches to correct PAL trace sequencing
The PAL trace buffer is stored as a circular buffer. Currently,
mapping the log to userspace starts at index zero to the max
buffer size, which does not output the log in correct time
sequence. Added patches to ensure the buffer starts at the lowest
timestamp and sequences in order.

The patch accounts for spurious events, which is observed when
a single event line is out of sequence but the event lines
proceeding it follows the sequence form the previous events.

Lastly, the patch accounts for collecting continuous dumps.
If an event is already dumped to userspace,the same event
will not be dumped again.

Change-Id: I7b1db0dabfde6b74b3290bcd4cc07c3e36986a42
Signed-off-by: Justin Park <quic_justpark@quicinc.com>
Signed-off-by: Debanjalee Roy <debanjal@qti.qualcomm.com>