mm-drivers: hw_fence: modify get_fence and import for reusable fences
Add support for synx_get_fence to return an error if hw-fence
is reusable. Add support for synt_import if hw-fence is reusable
but import flags are not reusable, or vice versa. Function will
return error in case of reusable fence is already signaled by
the time of function call.

Change-Id: I4440451a076f4afaf092682091356d4d745a4449
Signed-off-by: Aditya Shirsat <ashirsat@qti.qualcomm.com>