mm-drivers: hw_fence: clear fctl refcount regardless of signal_status
Currently, we skip clearing the fctl refcount when resetting the fence
if the fence is already signaled by hlos. This causes a fence leak in
case of synx_recover when dma-fence is already signaled for a given
fence, but fctl has yet to process the fence. Avoid this by ensuring
fctl_refcount is cleared regardless of signaling status.

Change-Id: I238a47b0b0324331b6aad76b94349edbfd3f2e3b
Signed-off-by: Grace An <gracan@qti.qualcomm.com>